+++ src/main/java/com/lokoko/domain/like/application/service/ProductLikeService.java
@@ -5,9 +5,9 @@ import com.lokoko.domain.like.domain.repository.ProductLikeRepository;
 import com.lokoko.domain.product.domain.entity.Product;
 import com.lokoko.domain.product.domain.repository.ProductRepository;
 import com.lokoko.domain.product.exception.ProductNotFoundException;
-import com.lokoko.domain.user.entity.User;
+import com.lokoko.domain.user.domain.entity.User;
 import com.lokoko.domain.user.exception.UserNotFoundException;
-import com.lokoko.domain.user.repository.UserRepository;
+import com.lokoko.domain.user.domain.repository.UserRepository;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
