+++ src/main/java/com/lokoko/domain/like/domain/entity/ProductLike.java
@@ -3,7 +3,7 @@ package com.lokoko.domain.like.domain.entity;
 import static jakarta.persistence.FetchType.LAZY;
 
 import com.lokoko.domain.product.domain.entity.Product;
-import com.lokoko.domain.user.entity.User;
+import com.lokoko.domain.user.domain.entity.User;
 import com.lokoko.global.common.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
