+++ src/main/java/com/lokoko/domain/product/application/crawler/NewProductCrawler.java
@@ -1,7 +1,7 @@
 package com.lokoko.domain.product.application.crawler;
 
-import com.lokoko.domain.image.entity.ProductImage;
-import com.lokoko.domain.image.repository.ProductImageRepository;
+import com.lokoko.domain.image.domain.entity.ProductImage;
+import com.lokoko.domain.image.domain.repository.ProductImageRepository;
 import com.lokoko.domain.product.domain.entity.Product;
 import com.lokoko.domain.product.domain.entity.enums.MainCategory;
 import com.lokoko.domain.product.domain.entity.enums.MiddleCategory;
