+++ src/main/java/com/lokoko/domain/product/domain/repository/ProductRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.lokoko.domain.product.domain.repository;
 
 
-import com.lokoko.domain.image.entity.QProductImage;
+import com.lokoko.domain.image.domain.entity.QProductImage;
 import com.lokoko.domain.like.domain.entity.QProductLike;
 import com.lokoko.domain.product.api.dto.NewProductProjection;
 import com.lokoko.domain.product.api.dto.PopularProductProjection;
